Addressing the Y2k38 issue for SAS tokens (#1807)
* Updating c-utility

* umock changes to allow uint64_t for SAS expiry time.

* Updating SAS internal API usage and UTs.

* Updating c-utility.

* Fixing UTs. Addressing PR feedback.

* Updating submodules.

+++ iothub_client/src/iothubtransport_mqtt_common.c
@@ -2641,7 +2641,7 @@ static int UpdateMqttConnectionStateIfNeeded(PMQTTTRANSPORT_HANDLE_DATA transpor
                 // If the credential type is not an x509 certificate then we shall renew the Sas_Token
                 if (cred_type != IOTHUB_CREDENTIAL_TYPE_X509 && cred_type != IOTHUB_CREDENTIAL_TYPE_X509_ECC)
                 {
-                    size_t sas_token_expiry = IoTHubClient_Auth_Get_SasToken_Expiry(transport_data->authorization_module);
+                    uint64_t sas_token_expiry = IoTHubClient_Auth_Get_SasToken_Expiry(transport_data->authorization_module);
                     if ((current_time - transport_data->mqtt_connect_time) / 1000 > (sas_token_expiry*SAS_REFRESH_MULTIPLIER))
                     {
                         /* Codes_SRS_IOTHUB_TRANSPORT_MQTT_COMMON_07_058: [ If the sas token has timed out IoTHubTransport_MQTT_Common_DoWork shall disconnect from the mqtt client and destroy the transport information and wait for reconnect. ] */
